+++ lib/libvmmapi/internal.h
@@ -7,11 +7,21 @@
 #ifndef __VMMAPI_INTERNAL_H__
 #define	__VMMAPI_INTERNAL_H__
 
-struct vmctx;
+struct vmctx {
+	int	fd;
+	uint32_t lowmem_limit;
+	int	memflags;
+	size_t	lowmem;
+	size_t	highmem;
+	char	*baseaddr;
+	char	*name;
+};
 
 struct vcpu {
 	struct vmctx *ctx;
 	int vcpuid;
 };
+
+int	vcpu_ioctl(struct vcpu *vcpu, u_long cmd, void *arg);
 
 #endif /* !__VMMAPI_INTERNAL_H__ */
